refactor: change name of homepage component to homepage-area
