+++ frontend/src/components/app/homepage/homepage-area.tsx
